% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_functions.R
\name{fdReconstruct}
\alias{fdReconstruct}
\title{Reconstruct imperfectly observed functional data}
\usage{
fdReconstruct(
  Y0.obs,
  Y1.obs = NULL,
  T.set = 1:dim(Y0.obs)[1],
  method = "CV",
  r = NULL,
  r.max = 20,
  w = NULL,
  pred.band = FALSE,
  p = 0.95
)
}
\arguments{
\item{Y0.obs}{Data matrix for target.}

\item{Y1.obs}{Data matrix for covariate (NULL for univariate case).}

\item{T.set}{Vector indicating functions to be reconstructed. By default,
all functions are reconstructed.}

\item{method}{Method for determining O-specific number of factors. Must be
either 'CV' for 5-fold cross validation or 'manual' for
manual choice of number of factors.}

\item{r}{Manual choice of O-specific number of factors (a vector).}

\item{r.max}{Maximum number of factors.}

\item{w}{Manual choice of weights (NULL uses integrated variances for
scaling).}

\item{pred.band}{Either TRUE or FALSE (default) indicating whether prediction
bands should be returned.}

\item{p}{Coverage probability for prediction bands (p = 0.95 by
default.)}
}
\value{
A list with the reconstructed data matrix 'X.hat', the
                 prediction bands 'U.hat' and 'L.hat' (only if
                 pred.band = TRUE) as well as vectors containing estimated
                 numbers of factors 'r.hat' and weights 'w.hat'.
}
\description{
Reconstruct imperfectly observed functional data
}
\examples{

set.seed(123)
data_comp <- GenObs(T = 50, complete = TRUE)
data_inco <- GenObs(T = 50, complete = FALSE)

Y0.obs <- rbind(data_inco$Y0.obs, data_comp$Y0.obs)
Y1.obs <- rbind(data_inco$Y1.obs, data_comp$Y1.obs)

reconst <- fdReconstruct(Y0.obs, Y1.obs, T.set = c(1), pred.band = TRUE)
plot(Y0.obs[1,], ylim = c(-0.5, 2.5))
lines(data_inco$X0[1,], lty = 2)
lines(reconst$X.hat[1,], col = "blue", lwd = 2)
lines(reconst$U.hat[1,], lty = 3)
lines(reconst$L.hat[1,], lty = 3)
}
