% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_num_factors.R
\name{RunCV}
\alias{RunCV}
\title{Run 5-fold cross validation for estimating the unknown rank.}
\usage{
RunCV(M.set = NULL, Y.obs, r.max = 20)
}
\arguments{
\item{M.set}{Set on which observations are missing. If M.set = NULL (i.e.
all observations are available) the rank is chosen such that
the fraction of variance explained exceeds 0.99.}

\item{Y.obs}{Data matrix.}

\item{r.max}{Maximum number of factors (r.max = 20 by default).}
}
\value{
Estimated number of factors.
}
\description{
Run 5-fold cross validation for estimating the unknown rank.
}
\examples{
set.seed(123)
data_comp <- GenObs(T = 50, N = 51, r = 4, complete = TRUE)
data_inco <- GenObs(T = 50, N = 51, r = 4, complete = FALSE)
Y0.obs <- rbind(data_inco$Y0.obs, data_comp$Y0.obs)
Y1.obs <- rbind(data_inco$Y1.obs, data_comp$Y1.obs)
RunCV(M.set = 1:10, cbind(Y0.obs, Y1.obs))
}
